refactor: combined page and event tracking into one new hook file use-posthog-tracking; updated references
