+++ src/components/case-study/case-study-image.tsx
@@ -7,7 +7,7 @@ import { getS3ImageUrl } from '@/lib/services/s3-service';
 import { logger } from '@/lib/utils/logger';
 import { LoadingSpinnerOverlay } from '@/components/ui/loading-spinner';
 import { useMotionVariant, useMotionTransition } from '@/lib/animations';
-import { useEventTracking } from '@/lib/hooks/use-event-tracking';
+import { useEventTracking } from '@/lib/hooks/use-posthog-tracking';
 
 const ImageLightbox = dynamic(() =>
   import('@/components/ui/image-lightbox').then((mod) => mod.ImageLightbox),
