+++ src/components/footer/legal-links.tsx
@@ -2,7 +2,7 @@
 
 import { Fragment } from 'react';
 import Link from 'next/link';
-import { useEventTracking } from '@/lib/hooks/use-event-tracking';
+import { useEventTracking } from '@/lib/hooks/use-posthog-tracking';
 
 interface LegalLinksProps {
   items: { name: string; href: string }[];
